+++ financial-connections/src/main/java/com/stripe/android/financialconnections/domain/GenerateFinancialConnectionsSessionManifest.kt
@@ -1,22 +1,22 @@
 package com.stripe.android.financialconnections.domain
 
-import com.stripe.android.financialconnections.model.LinkAccountSessionManifest
+import com.stripe.android.financialconnections.model.FinancialConnectionsSessionManifest
 import com.stripe.android.financialconnections.repository.FinancialConnectionsRepository
 import javax.inject.Inject
 
 /**
- * Fetches the [LinkAccountSessionManifest] from the Stripe API to get the hosted auth flow URL
+ * Fetches the [FinancialConnectionsSessionManifest] from the Stripe API to get the hosted auth flow URL
  * as well as the success and cancel callback URLs to verify.
  */
-internal class GenerateLinkAccountSessionManifest @Inject constructor(
+internal class GenerateFinancialConnectionsSessionManifest @Inject constructor(
     private val financialConnectionsRepository: FinancialConnectionsRepository,
 ) {
 
     suspend operator fun invoke(
         clientSecret: String,
         applicationId: String
-    ): LinkAccountSessionManifest {
-        return financialConnectionsRepository.generateLinkAccountSessionManifest(
+    ): FinancialConnectionsSessionManifest {
+        return financialConnectionsRepository.generateFinancialConnectionsSessionManifest(
             clientSecret = clientSecret,
             applicationId = applicationId
         )