+++ payments-core/src/main/java/com/stripe/android/model/FinancialConnectionsSession.kt
@@ -4,7 +4,7 @@ import com.stripe.android.core.model.StripeModel
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-internal data class BankConnectionsLinkedAccountSession(
+internal data class FinancialConnectionsSession(
     val clientSecret: String?,
     val id: String?,
 ) : StripeModel